REPL compilation: fix linenos of user code on py3.6
Commit 706674c7efb527f930072aaf896a402402c8c869 introduced a regression:
on python3.6, errors in usercode now start at line #2, not line #1.
This is due to a constant offset, -12, being passed to ast.increment_linenos.
When that commit added a `pass` line, the -12 should've turned into a -13, but that seems to have been forgotten.

Rather than just manually update the magic number, this commit computes it at runtime.