make ReplAsyncExecutor send sent messages back to the repl generator
Allows snippets such as:

```
m = yield '*smash* that mf like button'
await m.add_reaction('👍')
```

With this, a message would be sent, and a reaction added to that same message.
This required the addition of a dependency: async_generator. This library has a function called
yield_from_ which allows delegating to another async generator without a nasty while True loop, akin
to what `yield from` does in regular, synchronous generators. Generators created with this library are 100%
compatible with regular generators.

Doing so also made the `async for` loop in the jishaku_py command uglier: it was made into a `while True` loop
with two `except StopAsyncIteration: break` occurrences inside it. While this is uglier, this is the standard way
to commmunicate two-ways with an async generator, and is not that much of a hardship considering how little
ReplAsyncExecutor.traverse had to change.