Fix stdin to use a pty on Unix
This prevents the internal shell from trying to buffer stdin, causing
none of the data piped into it to actually go through until the process
is closed.

The holdover is only required on Unix. Windows doesn't support PTY but
it also doesn't seem to care and never buffers stdin as far as I can see
anyway.